backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.4.2
#openai>=1.6.0
pymupdf==1.23.4
pillow==10.0.1
python-dotenv==1.0.0
python-docx==0.8.11
pytest 